Tests: Rename setupDomEnv to useFakeDom and move to its own folder
