+++ test/test/helpers/use-fake-dom/index.js
